+++ python/acl_anthology/collections/collection.py
@@ -14,21 +14,15 @@
 
 from __future__ import annotations
 
-import sys
 from attrs import define, field, validators as v
 from lxml import etree
 from pathlib import Path
-from typing import Any, Iterator, Optional, cast, TYPE_CHECKING
-
-if sys.version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
+from typing import Any, Iterator, Optional, Self, cast, TYPE_CHECKING
 
 from ..containers import SlottedDict
 from ..exceptions import AnthologyDuplicateIDError, AnthologyInvalidIDError
 from ..text.markuptext import MarkupText
-from ..utils.attrs import auto_validate_types, int_to_str
+from ..utils.attrs import attach_custom_repr, auto_validate_types, int_to_str
 from ..utils.ids import infer_year, is_valid_collection_id
 from ..utils.logging import get_logger
 from ..utils import xml
@@ -46,6 +40,7 @@ if TYPE_CHECKING:
 log = get_logger()
 
 
+@attach_custom_repr
 @define(field_transformer=auto_validate_types)
 class Collection(SlottedDict[Volume]):
     """A collection of volumes and events, corresponding to an XML file in the `data/xml/` directory of the Anthology repo.
@@ -67,18 +62,19 @@ class Collection(SlottedDict[Volume]):
         raised_error_on_load: A flag indicating whether loading the XML file raised an error. If True, calling `load()` again won't do anything. This is to prevent the same exceptions being triggered over and over again by other functions accessing this collection (and thereby triggering a load) multiple times.
     """
 
-    id: str = field(converter=int_to_str)
+    id: str = field(converter=int_to_str)  # validator defined below
     parent: CollectionIndex = field(repr=False, eq=False)
     path: Path = field(converter=Path)
     event: Optional[Event] = field(
         init=False,
-        repr=False,
         default=None,
         validator=v.optional(v.instance_of(Event)),
     )
-    is_data_loaded: bool = field(init=False, repr=True, default=False)
-    is_modified: bool = field(init=False, repr=False, default=False)
-    raised_error_on_load: bool = field(init=False, repr=False, default=False)
+    is_data_loaded: bool = field(
+        init=False, default=False, metadata={"repr_omit_if": True}
+    )
+    is_modified: bool = field(init=False, default=False)
+    raised_error_on_load: bool = field(init=False, default=False)
 
     @id.validator
     def _check_id(self, _: Any, value: str) -> None: