+++ python/acl_anthology/people/person.py
@@ -16,12 +16,17 @@ from __future__ import annotations
 
 import attrs
 from attrs import define, field, setters
-from enum import Enum
+from enum import StrEnum
 from typing import Any, Iterator, Optional, Sequence, TYPE_CHECKING
-import warnings
+import sys
+
+if sys.version_info >= (3, 13):
+    from warnings import deprecated
+else:
+    from typing_extensions import deprecated
 
 from ..exceptions import AnthologyException, AnthologyInvalidIDError
-from ..utils.attrs import auto_validate_types
+from ..utils.attrs import attach_custom_repr, auto_validate_types, repr_item_ids
 from ..utils.ids import (
     AnthologyID,
     AnthologyIDTuple,
@@ -34,12 +39,12 @@ from ..utils.ids import (
 from . import Name
 
 if TYPE_CHECKING:
-    from . import NameSpecification
+    from . import NameSpecification, PersonIndex
     from ..anthology import Anthology
     from ..collections import Paper, Volume
 
 
-class NameLink(Enum):
+class NameLink(StrEnum):
     """How a Name was connected to a Person."""
 
     EXPLICIT = "explicit"
@@ -47,6 +52,9 @@ class NameLink(Enum):
 
     INFERRED = "inferred"
     """Name was connected to this Person via slug matching heuristic."""
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}.{self.name}"
 
 
 def _name_list_converter(
@@ -77,7 +85,7 @@ def _update_person_index(person: Person, attr: attrs.Attribute[Any], value: str)
 
     Intended to be called from `on_setattr` of an [attrs.field][].
     """
-    index = person.parent.people
+    index = person.parent
     if attr.name == "id":
         index._update_id(person.id, value)
     elif attr.name == "orcid":
@@ -85,6 +93,7 @@ def _update_person_index(person: Person, attr: attrs.Attribute[Any], value: str)
     return value
 
 
+@attach_custom_repr
 @define(field_transformer=auto_validate_types)
 class Person:
     """A natural person.
@@ -106,14 +115,17 @@ class Person:
         is_explicit: If True, this person's ID is explicitly defined in `people.yaml`.  You probably want to use [`make_explicit()`][acl_anthology.people.person.Person.make_explicit] rather than change this attribute.
     """
 
-    id: str = field(on_setattr=[setters.validate, _update_person_index])
-    parent: Anthology = field(repr=False, eq=False)
+    id: str = field(
+        on_setattr=[setters.validate, _update_person_index],
+        metadata={"repr_omits_field_name": True},
+    )
+    parent: PersonIndex = field(repr=False, eq=False)
     _names: list[tuple[Name, NameLink]] = field(
-        factory=list, converter=_name_list_converter
+        factory=list,
+        converter=_name_list_converter,
+        metadata={"repr_omits_field_name": True},
     )
-    item_ids: list[AnthologyIDTuple] = field(
-        factory=list, repr=lambda x: f"<list of {len(x)} AnthologyIDTuple objects>"
-    )
+    item_ids: list[AnthologyIDTuple] = field(factory=list, repr=repr_item_ids)
     orcid: Optional[str] = field(
         default=None,
         on_setattr=[_orcid_converter_and_validator, _update_person_index],
@@ -122,7 +134,7 @@ class Person:
     degree: Optional[str] = field(default=None)
     similar_ids: list[str] = field(factory=list)
     disable_name_matching: Optional[bool] = field(default=False, converter=bool)
-    is_explicit: Optional[bool] = field(default=False, converter=bool)
+    is_explicit: Optional[bool] = field(default=False, converter=bool, repr=False)
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, Person):
@@ -133,6 +145,11 @@ class Person:
         return hash(self.id)
 
     @property
+    def root(self) -> Anthology:
+        """The Anthology instance to which this object belongs."""
+        return self.parent.parent
+
+    @property
     def names(self) -> list[Name]:
         """A list of all names associated with this person."""
         return [name for (name, _) in self._names]
@@ -140,9 +157,9 @@ class Person:
     @names.setter
     def names(self, values: list[Name]) -> None:
         for name, _ in self._names:
-            self.parent.people._remove_name(self.id, name)
+            self.parent._remove_name(self.id, name)
         for name in values:
-            self.parent.people._add_name(self.id, name)
+            self.parent._add_name(self.id, name)
         self._names = _name_list_converter(values)
 
     @property
@@ -184,7 +201,7 @@ class Person:
         link_type = NameLink.INFERRED if inferred else NameLink.EXPLICIT
         if not self.has_name(name):
             self._names.append((name, link_type))
-            self.parent.people._add_name(self.id, name)
+            self.parent._add_name(self.id, name)
         elif (name, link_type) not in self._names:
             # ensure that name is re-inserted at same position
             idx = self.names.index(name)
@@ -204,7 +221,7 @@ class Person:
             ValueError: If this name was not explicitly linked to this person.
         """
         self._names.remove((name, NameLink.EXPLICIT))
-        self.parent.people._remove_name(self.id, name)
+        self.parent._remove_name(self.id, name)
 
     def has_name(self, name: Name) -> bool:
         """
@@ -228,7 +245,7 @@ class Person:
             AnthologyException: If `self.explicit` is False.
             AnthologyInvalidIDError: If the supplied ID is not well-formed, or if it already exists in the PersonIndex.
         """
-        if new_id in self.parent.people:
+        if new_id in self.parent:
             exc = AnthologyInvalidIDError(new_id, f"Person ID already exists: {new_id}")
             exc.add_note("Did you want to use merge_into() instead?")
             raise exc
@@ -265,12 +282,12 @@ class Person:
         if self.is_explicit:
             raise AnthologyException(f"Person '{self.id}' is already explicit")
         if new_id is None:
-            new_id = self.parent.people.generate_person_id(self)
+            new_id = self.parent.generate_person_id(self)
         elif not is_verified_person_id(new_id):
             raise AnthologyInvalidIDError(
                 new_id, f"Not a valid verified-person ID: {new_id}"
             )
-        elif new_id in self.parent.people:
+        elif new_id in self.parent:
             raise AnthologyException(f"ID already exists in the index: {new_id}")
 
         self.is_explicit = True
@@ -279,12 +296,10 @@ class Person:
         self.id = new_id  # triggers update in PersonIndex
         self._names = [(name, NameLink.EXPLICIT) for name, _ in self._names]
 
+    @deprecated(
+        "Person.merge_with_explicit() is deprecated in favor of Person.merge_into()"
+    )
     def merge_with_explicit(self, person: Person) -> None:  # pragma: no cover
-        warnings.warn(
-            DeprecationWarning(
-                "Person.merge_with_explicit() is deprecated in favor of Person.merge_into()"
-            )
-        )
         self.merge_into(person)
 
     def merge_into(self, other: Person) -> None:
@@ -358,7 +373,7 @@ class Person:
         """Returns an iterator over all Anthology items associated with this person, regardless of their type."""
         # TODO: This does not consider talks yet!
         for anthology_id in self.item_ids:
-            item = self.parent.get(anthology_id)
+            item = self.root.get(anthology_id)
             if item is None:
                 raise ValueError(
                     f"Person {self.id} lists associated item {build_id_from_tuple(anthology_id)}, which doesn't exist"
@@ -380,7 +395,7 @@ class Person:
         for anthology_id in self.item_ids:
             paper_id = anthology_id[-1]
             if paper_id is not None:
-                paper = self.parent.get_paper(anthology_id)
+                paper = self.root.get_paper(anthology_id)
                 if paper is None:
                     raise ValueError(
                         f"Person {self.id} lists associated paper {build_id_from_tuple(anthology_id)}, which doesn't exist"
@@ -392,7 +407,7 @@ class Person:
         for anthology_id in self.item_ids:
             paper_id = anthology_id[-1]
             if paper_id is None:
-                volume = self.parent.get_volume(anthology_id)
+                volume = self.root.get_volume(anthology_id)
                 if volume is None:
                     raise ValueError(
                         f"Person {self.id} lists associated volume {build_id_from_tuple(anthology_id)}, which doesn't exist"