+++ python/tests/collections/volume_test.py
@@ -169,7 +169,7 @@ def test_volume_attributes_2022acl_long(anthology):
     assert volume.year == "2022"
     assert volume.pdf.name == "2022.acl-long"
     assert volume.pdf.checksum == "b8317652"
-    assert volume.venue_ids == ["acl"]
+    assert volume.venue_ids == ("acl",)
     assert volume.venue_acronym == "ACL"
     assert not volume.is_workshop
     assert isinstance(volume.frontmatter, Paper) and volume.frontmatter.id == "0"
@@ -188,7 +188,7 @@ def test_volume_attributes_2022acl_demo(anthology):
     assert volume.year == "2022"
     assert volume.pdf.name == "2022.acl-demo"
     assert volume.pdf.checksum == "d92e3f4d"
-    assert volume.venue_ids == ["acl"]
+    assert volume.venue_ids == ("acl",)
     assert volume.venue_acronym == "ACL"
     assert not volume.is_workshop
     assert isinstance(volume.frontmatter, Paper) and volume.frontmatter.id == "0"
@@ -200,7 +200,7 @@ def test_volume_attributes_j89(anthology):
     volume = anthology.get_volume("J89-1")
     assert isinstance(volume, Volume)
     assert volume.id == "1"
-    assert volume.venue_ids == ["cl"]
+    assert volume.venue_ids == ("cl",)
     assert volume.venue_acronym == "CL"
     assert volume.year == "1989"
     assert not volume.is_workshop
@@ -217,7 +217,7 @@ def test_volume_attributes_naloma(anthology):
     assert volume.id == "1"
     assert volume.year == "2022"
     assert volume.is_workshop
-    assert volume.venue_ids == ["nlma", "ws"]
+    assert volume.venue_ids == ("nlma", "ws")
     assert volume.venue_acronym == "NALOMA"
     assert isinstance(volume.frontmatter, Paper) and volume.frontmatter.id == "0"
 
@@ -267,7 +267,7 @@ def test_paper_setattr_on_namespec_sets_collection_is_modified(anthology):
 
 def test_volume_venues_j89(anthology):
     volume = anthology.get_volume("J89-1")
-    assert volume.venue_ids == ["cl"]
+    assert volume.venue_ids == ("cl",)
     venues = volume.venues()
     assert len(venues) == 1
     assert venues[0].id == "cl"
@@ -275,7 +275,7 @@ def test_volume_venues_j89(anthology):
 
 def test_volume_venues_naloma(anthology):
     volume = anthology.get_volume("2022.naloma-1")
-    assert volume.venue_ids == ["nlma", "ws"]
+    assert volume.venue_ids == ("nlma", "ws")
     venues = volume.venues()
     assert len(venues) == 2
     assert venues[0].id == "nlma"
@@ -311,7 +311,7 @@ def test_volume_with_multiple_venues(anthology):
     )
     assert volume.full_id == "2092.acl-1"
     assert volume.title == volume_title
-    assert volume.venue_ids == ["acl", "lrec"]
+    assert volume.venue_ids == ("acl", "lrec")
     assert volume.venue_acronym == "ACL-LREC"
 
 
@@ -373,7 +373,7 @@ def test_volume_generate_paper_id(anthology):
 def test_volume_create_paper_implicit(anthology):
     volume = anthology.get_volume("2022.acl-long")
     assert not volume.collection.is_modified
-    authors = [NameSpec("Bollmann, Marcel")]
+    authors = (NameSpec("Bollmann, Marcel"),)
     paper = volume.create_paper(
         title="The awesome paper I have never written",
         authors=authors,
@@ -395,7 +395,7 @@ def test_volume_create_paper_implicit(anthology):
 def test_volume_create_paper_explicit(anthology):
     volume = anthology.get_volume("2022.acl-long")
     assert not volume.collection.is_modified
-    authors = [NameSpec("Bollmann, Marcel")]
+    authors = (NameSpec("Bollmann, Marcel"),)
     paper = volume.create_paper(
         title="The awesome paper I have never written",
         authors=authors,
@@ -416,7 +416,7 @@ def test_volume_create_paper_explicit(anthology):
 
 def test_volume_create_paper_with_duplicate_id_should_fail(anthology):
     volume = anthology.get_volume("2022.acl-long")
-    authors = [NameSpec("Bollmann, Marcel")]
+    authors = (NameSpec("Bollmann, Marcel"),)
     with pytest.raises(ValueError):
         _ = volume.create_paper(
             title="The awesome paper I have never written",
@@ -439,7 +439,7 @@ def test_volume_create_paper_with_editors(anthology):
     volume = anthology.get_volume("2022.acl-long")
 
     # For most papers, the editors are the volume's editors
-    authors = [NameSpec("Bollmann, Marcel")]
+    authors = (NameSpec("Bollmann, Marcel"),)
     paper = volume.create_paper(
         title="The awesome paper I have never written",
         authors=authors,
@@ -448,7 +448,7 @@ def test_volume_create_paper_with_editors(anthology):
     assert paper.get_editors() == volume.editors
 
     # But the schema allows paper-level editors too
-    editors = [NameSpec("Calzolari, Nicoletta")]
+    editors = (NameSpec("Calzolari, Nicoletta"),)
     paper = volume.create_paper(
         title="The awesome paper I have never written",
         authors=authors,
@@ -465,7 +465,7 @@ def test_volume_create_paper_should_update_person(anthology, pre_load):
         anthology.people.load()  # otherwise we test creation, not updating
 
     volume = anthology.get_volume("2022.acl-long")
-    authors = [NameSpec("Berg-Kirkpatrick, Taylor")]
+    authors = (NameSpec("Berg-Kirkpatrick, Taylor"),)
     paper = volume.create_paper(
         title="The awesome paper I have never written",
         authors=authors,
@@ -484,7 +484,7 @@ def test_volume_create_paper_should_update_personindex(anthology, pre_load):
         anthology.people.load()  # otherwise we test creation, not updating
 
     volume = anthology.get_volume("2022.acl-long")
-    authors = [NameSpec("Nonexistant, Guy Absolutely")]
+    authors = (NameSpec("Nonexistant, Guy Absolutely"),)
     paper = volume.create_paper(
         title="An entirely imaginary paper",
         authors=authors,
@@ -521,7 +521,7 @@ def test_volume_add_editor(anthology):
     assert volume.full_id_tuple not in person.item_ids
 
     # Adding this editor to the volume
-    volume.editors += [ns]
+    volume.editors += (ns,)
     # Person should be updated after resetting indices
     anthology.reset_indices()
     person = ns.resolve()